Moved getBody function to utility.js
